Fixes #1160 + A bit of testing code cleanup:
* Reversed include orders - see issue #681; this exposed some missing includes, so:
* Added missing includes, especially for standard library headers, to some test programs (and some non-test utility headers)
* Moved some non-inlined, non-template test utility code from headers into actual source files (so that it doesn't get defined twice)
* Replaced `types.h` uses of `stdint.h` types with the raw types, under the same assumptions as for `gdf_date32`, `gdf_date64` etc.
* Minor changes to the column printing code
