+++ cpp/src/utilities/type_dispatcher.hpp
@@ -16,10 +16,13 @@
 #ifndef TYPE_DISPATCHER_HPP
 #define TYPE_DISPATCHER_HPP
 
-#include "NVStrings.h"
-#include "cudf/types.h"
 #include "wrapper_types.hpp"
 #include "release_assert.cuh"
+
+#include <cudf/types.h>
+
+#include <NVStrings.h>
+
 #include <cassert>
 #include <utility>
