+++ cpp/tests/bitmask/valid_ops_test.cu
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-#include "gtest/gtest.h"
-#include "gmock/gmock.h"
+#include <tests/utilities/cudf_test_utils.cuh>
+#include <tests/utilities/cudf_test_fixtures.h>
 
 #include <cudf.h>
-#include <cudf/functions.h>
 
-#include "cuda_profiler_api.h"
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
 
-#include "tests/utilities/cudf_test_utils.cuh"
-#include "tests/utilities/cudf_test_fixtures.h"
+#include <cuda_profiler_api.h>
 
 #include <chrono>
