+++ cpp/tests/filter/helper/utils.cu
@@ -1,11 +1,7 @@
 
-#include <iostream>
-#include <cudf.h>
-#include <cudf/functions.h>
-#include <cuda_runtime.h>
-#include <limits.h>
-#include <gtest/gtest.h>
 #include "utils.cuh"
+
+#include <climits>
 
 
 gdf_valid_type * get_gdf_valid_from_device(gdf_column* column) {
