+++ cpp/tests/filter/helper/utils.cuh
@@ -2,19 +2,22 @@
 #ifndef GDF_TEST_UTILS
 #define GDF_TEST_UTILS
 
-#include <iostream>
 #include <cudf.h>
-#include <cudf/functions.h>
+
 #include <thrust/functional.h>
 #include <thrust/device_ptr.h>
 #include <thrust/iterator/counting_iterator.h>
-#include "utilities/type_dispatcher.hpp"
+#include <utilities/type_dispatcher.hpp>
 
+#include <rmm/rmm.h>
+
+#include <gtest/gtest.h>
+
+#include <iostream>
 #include <string>
 #include <functional>
 #include <vector>
 #include <tuple>
-#include <rmm/rmm.h>
 
 inline auto get_number_of_bytes_for_valid (size_t column_size) -> size_t {
     return sizeof(gdf_valid_type) * (column_size + GDF_VALID_BITSIZE - 1) / GDF_VALID_BITSIZE;
