+++ cpp/tests/filter/test_example.cu
@@ -15,20 +15,22 @@
  * limitations under the License.
  */
 
-#include "gtest/gtest.h"
-
-#include <iostream>
-#include <thrust/functional.h>
-#include <thrust/device_ptr.h>
-
-#include <thrust/execution_policy.h>
-#include <cuda_runtime.h>
 #include "helper/utils.cuh"
 
-#include <cudf.h>
-#include <cudf/functions.h>
+#include <tests/utilities/cudf_test_fixtures.h>
 
-#include "tests/utilities/cudf_test_fixtures.h"
+#include <cudf.h>
+
+#include <thrust/functional.h>
+#include <thrust/device_ptr.h>
+#include <thrust/execution_policy.h>
+
+#include <gtest/gtest.h>
+
+#include <cuda_runtime.h>
+
+#include <iostream>
+
 struct Example : public GdfTest {};
 
 /*
