+++ cpp/tests/filter/test_filter_ops.cu
@@ -16,17 +16,22 @@
  * limitations under the License.
  */
 
-#include "gtest/gtest.h"
-#include <iostream>
+#include "helper/utils.cuh"
+
+#include <tests/utilities/cudf_test_fixtures.h>
+
 #include <cudf.h>
-#include <cudf/functions.h>
+
 #include <thrust/functional.h>
 #include <thrust/device_ptr.h>
 #include <thrust/execution_policy.h>
+
+#include <gtest/gtest.h>
+
 #include <cuda_runtime.h>
+
+#include <iostream>
 #include <tuple>
-#include "helper/utils.cuh"
-#include "tests/utilities/cudf_test_fixtures.h"
 
 /*
  ============================================================================
