+++ cpp/tests/groupby/groupby_test.cu
@@ -13,29 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <cstdlib>
-#include <iostream>
-#include <vector>
-#include <map>
-#include <utility>
-#include <type_traits>
-#include <typeinfo>
-#include <memory>
-
-#include "gtest/gtest.h"
-#include "gmock/gmock.h"
-
-#include <cudf.h>
-#include <cudf/functions.h>
-
-#include "utilities/cudf_utils.h"
-
-#include <tests/utilities/cudf_test_fixtures.h>
 
 // See this header for all of the recursive handling of tuples of vectors
 #include "test_parameters.cuh"
 #include "groupby_test_helpers.cuh"
-#include <bitmask/legacy_bitmask.hpp>
+
+#include <tests/utilities/cudf_test_fixtures.h>
+
+#include <utilities/cudf_utils.h>
+
+#include <cudf.h>
+
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
+
+#include <iostream>
+#include <vector>
+#include <utility>
+#include <type_traits>
+#include <typeinfo>
+#include <memory>
+#include <cstdlib>
+
 
 // A new instance of this class will be created for each *TEST(GroupTest, ...)
 // Put all repeated setup and validation stuff here
