+++ cpp/tests/types/types_test.cpp
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <random>
-#include "gtest/gtest.h"
 
 #include <utilities/wrapper_types.hpp>
-
 #include <cudf.h>
 
+#include <gtest/gtest.h>
+
+#include <random>
 
 
 /**
