+++ cpp/tests/utilities/cudf_test_fixtures.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gtest/gtest.h"
+#include <gtest/gtest.h>
 
 #include <rmm/rmm.h>
