+++ cpp/tests/utilities/tuple_vectors.h
@@ -16,10 +16,15 @@
 #ifndef TUPLE_VECTORS_H
 #define TUPLE_VECTORS_H
 
+// See this header for all of the handling of valids' vectors
+// #include <tests/utilities/valid_vectors.h>
+
 #include <vector>
 #include <type_traits>
 #include <iostream>
 #include <cstdlib>
+#include <algorithm>
+#include <ostream>
 #include <iterator>
 
 template <typename... T>
