+++ cpp/tests/utilities_tests/column_wrapper_tests.cu
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-#include "tests/utilities/column_wrapper.cuh"
-#include "tests/utilities/cudf_test_fixtures.h"
-#include "utilities/type_dispatcher.hpp"
-#include "utilities/wrapper_types.hpp"
+#include <tests/utilities/column_wrapper.cuh>
+#include <tests/utilities/cudf_test_fixtures.h>
+#include <utilities/type_dispatcher.hpp>
+#include <utilities/wrapper_types.hpp>
 
-#include "gmock/gmock.h"
-#include "gtest/gtest.h"
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
 
 #include <bitset>
 #include <cstdint>
